requirements.txt
streamlit
google-generativeai
PyMuPDF
pandas
streamlit-option-menu